refactor: share runtime compatibility check
