+++ lambda/shared/runtime-compatibility.js
@@ -1,0 +1,12 @@
+export const supportsCompatibilityVersion = (candidate, current) => {
+  const candidateVersion = Number(candidate);
+  const currentVersion = Number(current);
+  return (
+    Number.isInteger(candidateVersion) &&
+    Number.isInteger(currentVersion) &&
+    candidateVersion >= currentVersion - 1 &&
+    candidateVersion <= currentVersion
+  );
+};
+
+export default { supportsCompatibilityVersion };
